requirements.txt
python-telegram-bot~=13.15
environs~=9.5.0
redis~=4.5.5